Extract parsing routine for groups of data separated by a newline character. This was needed initially for day 4, but we can reuse a generalized approach in day 6.
In the process of refactoring, we ran into a (not-so?) common bug with adding collections
to collections. Note that the CoC ends up just holding on to /references/. The solution
here is to make a deep copy each time we insert into the CoC.

Am I getting addicted to Java streams? I don't even fully understand them.
I find myself typing out a lot of code then replacing them with chained calls
on a stream object (e.g. map, filter, etc.). Is this healthy?