Solution to day 15, part 1
Started with map-based implementation that didn't work for example test cases.

Then started with list-based impl that didn't work either, until I realized that
I made a typo on the expected value in a unit test assert. Turns out the
map-based impl works for the simple tests.

Map-based impl is actually buggy with the rest of the part 1 tests. Will fix now.